Lights (#115)
* Added lighting system. Fixed non-invertible camera matrices, texture wrapping, and some data conversion warnings.

* Deserialized Enabled and Layer properties of lighting system.

* Added serialization and inspection of light properties, doxygenated, ready to merge

* Fixed light glitches, added lights to win and lose scenes.

* Fixed reading layers

* fixed bug where adding components to Entities not in the scene would call their OnInit functions

* Based LightingSystem on ComponentSystem. Fixed flickering when light sources are removed from scene.

---------

+++ Source/Pathfinder.cpp
@@ -334,7 +334,7 @@ bool Pathfinder::IsWalkable(glm::vec2 pos) const
     /// @brief  inspector for choosing which tiles are walkable
     void Pathfinder::inspectWalkables()
     {
-        if ( ImGui::BeginListBox( "walkable tiles", ImVec2( 82, (int)m_Walkables.size() * 24 + 26 ) ) == false )
+        if ( ImGui::BeginListBox( "walkable tiles", ImVec2( 82.0f, (float)m_Walkables.size() * 24.0f + 26.0f ) ) == false )
         {
             return;
         }
